Move `TypeSystem` to `mc-java-base`
