+++ mc-java-rejection/src/main/kotlin/io/spine/tools/mc/java/rejection/RThrowableCode.kt
@@ -36,6 +36,7 @@ import io.spine.tools.java.code.GeneratedBy
 import io.spine.tools.java.code.field.FieldName
 import io.spine.tools.java.constructorSpec
 import io.spine.tools.java.methodSpec
+import io.spine.tools.mc.java.TypeSystem
 import io.spine.tools.mc.java.rejection.Javadoc.forConstructorOfThrowable
 import io.spine.tools.mc.java.rejection.Javadoc.forThrowableOf
 import javax.lang.model.element.Modifier.FINAL
